Remove duplicate code from getBoostedPoolStateWithBalancesV3
