+++ src/entities/utils/getPoolStateWithBalancesV3.ts
@@ -1,17 +1,144 @@
-import { createPublicClient, formatEther, formatUnits, http } from 'viem';
+import {
+    createPublicClient,
+    erc4626Abi,
+    formatEther,
+    formatUnits,
+    http,
+    PublicClient,
+} from 'viem';
 
 import { HumanAmount } from '@/data';
 import { CHAINS, VAULT_V3 } from '@/utils';
 
 import { getSortedTokens } from './getSortedTokens';
-import { PoolState, PoolStateWithBalances } from '../types';
+import {
+    PoolState,
+    PoolStateWithBalances,
+    PoolStateWithUnderlyingBalances,
+    PoolStateWithUnderlyings,
+    PoolTokenWithUnderlying,
+} from '../types';
 import { vaultExtensionAbi_V3 } from '@/abi';
+import { TokenAmount } from '../tokenAmount';
 
 export const getPoolStateWithBalancesV3 = async (
     poolState: PoolState,
     chainId: number,
     rpcUrl: string,
 ): Promise<PoolStateWithBalances> => {
+    const publicClient = createPublicClient({
+        transport: http(rpcUrl),
+        chain: CHAINS[chainId],
+    });
+
+    const { tokenAmounts, totalShares } = await getTokenAmountsAndTotalShares(
+        chainId,
+        poolState,
+        publicClient,
+    );
+
+    const poolStateWithBalances: PoolStateWithBalances = {
+        ...poolState,
+        tokens: tokenAmounts.map((tokenAmount, i) => ({
+            address: tokenAmount.token.address,
+            decimals: tokenAmount.token.decimals,
+            index: i,
+            balance: formatUnits(
+                tokenAmount.amount[i],
+                tokenAmount.token.decimals,
+            ) as HumanAmount,
+        })),
+        totalShares: formatEther(totalShares) as HumanAmount,
+    };
+    return poolStateWithBalances;
+};
+
+export const getBoostedPoolStateWithBalancesV3 = async (
+    poolState: PoolStateWithUnderlyings,
+    chainId: number,
+    rpcUrl: string,
+): Promise<PoolStateWithUnderlyingBalances> => {
+    const publicClient = createPublicClient({
+        transport: http(rpcUrl),
+        chain: CHAINS[chainId],
+    });
+
+    const { tokenAmounts, totalShares } = await getTokenAmountsAndTotalShares(
+        chainId,
+        poolState,
+        publicClient,
+    );
+
+    const sortedTokens = [...poolState.tokens].sort(
+        (a, b) => a.index - b.index,
+    );
+
+    const underlyingBalances = await getUnderlyingBalances(
+        sortedTokens,
+        tokenAmounts,
+        publicClient,
+    );
+
+    const poolStateWithUnderlyingBalances: PoolStateWithUnderlyingBalances = {
+        ...poolState,
+        tokens: sortedTokens.map((token, i) => ({
+            ...token,
+            balance: formatUnits(
+                tokenAmounts[i].amount,
+                token.decimals,
+            ) as HumanAmount,
+            underlyingToken:
+                token.underlyingToken === null
+                    ? null
+                    : {
+                          ...token.underlyingToken,
+                          balance: formatUnits(
+                              underlyingBalances.shift() as bigint,
+                              token.underlyingToken.decimals,
+                          ) as HumanAmount,
+                      },
+        })),
+        totalShares: formatEther(totalShares) as HumanAmount,
+    };
+    return poolStateWithUnderlyingBalances;
+};
+
+const getUnderlyingBalances = async (
+    sortedTokens: PoolTokenWithUnderlying[],
+    tokenAmounts: TokenAmount[],
+    publicClient: PublicClient,
+) => {
+    const getUnderlyingBalancesContracts = sortedTokens
+        .filter((token) => token.underlyingToken !== null)
+        .map((token, i) => ({
+            address: token.address,
+            abi: erc4626Abi,
+            functionName: 'previewRedeem' as const,
+            args: [tokenAmounts[i].amount] as const,
+        }));
+
+    const underlyingBalanceOutputs = await publicClient.multicall({
+        contracts: [...getUnderlyingBalancesContracts],
+    });
+    if (
+        underlyingBalanceOutputs.some((output) => output.status === 'failure')
+    ) {
+        throw new Error(
+            'Error: Unable to get underlying balances for v3 pool.',
+        );
+    }
+
+    const underlyingBalances = underlyingBalanceOutputs.map(
+        (output) => output.result as bigint,
+    );
+    return underlyingBalances;
+};
+
+const getTokenAmountsAndTotalShares = async (
+    chainId: number,
+    poolState: PoolState,
+    publicClient: PublicClient,
+) => {
     const totalSupplyContract = {
         address: VAULT_V3[chainId],
         abi: vaultExtensionAbi_V3,
@@ -25,10 +152,6 @@ export const getPoolStateWithBalancesV3 = async (
         args: [poolState.address] as const,
     };
 
-    const publicClient = createPublicClient({
-        transport: http(rpcUrl),
-        chain: CHAINS[chainId],
-    });
     const outputs = await publicClient.multicall({
         contracts: [totalSupplyContract, getBalanceContracts],
     });
@@ -40,19 +163,12 @@ export const getPoolStateWithBalancesV3 = async (
     }
 
     const totalShares = outputs[0].result as bigint;
-    const balances = outputs[1].result as bigint[];
+    const balancesScale18 = outputs[1].result as bigint[];
 
-    const sortedTokens = getSortedTokens(poolState.tokens, chainId);
+    const poolTokens = getSortedTokens(poolState.tokens, chainId);
+    const tokenAmounts = poolTokens.map((token, i) =>
+        TokenAmount.fromScale18Amount(token, balancesScale18[i]),
+    );
 
-    const poolStateWithBalances: PoolStateWithBalances = {
-        ...poolState,
-        tokens: sortedTokens.map((token, i) => ({
-            address: token.address,
-            decimals: token.decimals,
-            index: i,
-            balance: formatUnits(balances[i], token.decimals) as HumanAmount,
-        })),
-        totalShares: formatEther(totalShares) as HumanAmount,
-    };
-    return poolStateWithBalances;
+    return { tokenAmounts, totalShares };
 };
